requirements.txt
Flask==3.0.3
minio==7.2.7
peewee==3.17.5
Pillow==10.3.0
PyJWT==2.8.0
PyYAML==6.0.1
Requests==2.32.3
Werkzeug==3.0.3
